KAFKA-16593: Migrate ConsumerRunnable to ConsumerGroupExecutor as a static inner class
